feat: init 파일을 만들어서 따로 init에 관련한 함수 로직 분리
