+++ src/js/modules/CarName.ts
@@ -1,15 +1,10 @@
 import { carNamesSection, carNameDiv } from './@share/view.js';
 import { ERROR_MESSAGE } from './@share/message.js';
+import { carNameInputInit } from './@share/init.js';
 
 const CarNameComponent = ({ $app, carNames }: { $app: HTMLDivElement | null; carNames: string }): void => {
   const checkCarNames = (carNameList: string[]) =>
     carNameList.length === carNameList.filter((x) => x.length <= 5 && x !== '').length;
-
-  const carNameInputInit = () => {
-    const carNameInput = document.getElementById('car-name-input') as HTMLInputElement;
-    carNameInput.value = '';
-    carNameInput.focus();
-  };
 
   const render = (JSX: string): void => {
     const sectionElement = $app;
